Updated UsersController and added JwtServices to Business Layer
