+++ BusinessLogicLayer/DtoModels/Authorization/TokenCredentials.cs
@@ -1,11 +1,10 @@
 ﻿using System;
 
-namespace SAPex.Models.Authorization
+namespace BusinessLogicLayer.DtoModels.Authorization
 {
     public class TokenCredentials
     {
         public string AccessToken { get; set; }
-
         public string RefreshToken { get; set; }
     }
 }
