+++ BusinessLogicLayer/DtoModels/Authorization/UserCredentials.cs
@@ -1,11 +1,10 @@
 ﻿using System;
 
-namespace SAPex.Models.Authorization
+namespace BusinessLogicLayer.DtoModels.Authorization
 {
     public class UserCredentials
     {
         public string Email { get; set; }
-
         public string Password { get; set; }
     }
 }
