+++ SAPex/Controllers/UsersController.cs
@@ -1,0 +1,25 @@
+﻿using BusinessLogicLayer.Services._Temp;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+
+namespace SAPex.Controllers
+{
+    [Route("api/users")]
+    [ApiController]
+    [Authorize]
+    public class UsersController : ControllerBase
+    {
+        private readonly UserService _userService;
+
+        public UsersController(UserService userService)
+        {
+            _userService = userService;
+        }
+
+        [HttpGet]
+        public IActionResult Get()
+        {
+            return Ok(_userService.FindAll());
+        }
+    }
+}
